requirements.txt
dnspython==2.2.1
motor==3.0.0
pycryptodome==3.17
orjson